+++ src/main/java/kitchenpos/domain/menu/MenuProduct.java
@@ -1,5 +1,7 @@
 package kitchenpos.domain.menu;
 
+import kitchenpos.domain.product.Product;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
