+++ src/main/java/kitchenpos/domain/order/Order.java
@@ -1,5 +1,7 @@
 package kitchenpos.domain.order;
 
+import kitchenpos.domain.table.OrderTable;
+
 import javax.persistence.Column;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
