+++ src/main/java/kitchenpos/domain/order/service/dto/TableGroupResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.domain.order.service.dto;
 
-import kitchenpos.domain.order.TableGroup;
+import kitchenpos.domain.table.TableGroup;
 
 import java.time.LocalDateTime;
 import java.util.List;
