+++ src/main/java/kitchenpos/domain/product/repository/ProductRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.domain.menu.repository;
+package kitchenpos.domain.product.repository;
 
-import kitchenpos.domain.menu.Product;
+import kitchenpos.domain.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
