+++ src/main/java/kitchenpos/domain/table/OrderTables.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.order;
+package kitchenpos.domain.table;
 
 import org.springframework.util.CollectionUtils;
