+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -3,14 +3,14 @@ package kitchenpos.application;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItems;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order.OrderTable;
-import kitchenpos.domain.order.TableGroup;
 import kitchenpos.domain.order.repository.OrderRepository;
-import kitchenpos.domain.order.repository.OrderTableRepository;
-import kitchenpos.domain.order.repository.TableGroupRepository;
-import kitchenpos.domain.order.service.TableGroupService;
 import kitchenpos.domain.order.service.dto.TableGroupCreateRequest;
 import kitchenpos.domain.order.service.dto.TableGroupResponse;
+import kitchenpos.domain.table.OrderTable;
+import kitchenpos.domain.table.TableGroup;
+import kitchenpos.domain.table.repository.OrderTableRepository;
+import kitchenpos.domain.table.repository.TableGroupRepository;
+import kitchenpos.domain.table.service.TableGroupService;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
