+++ src/test/java/kitchenpos/application/fixture/OrderFixture.java
@@ -4,7 +4,7 @@ import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.order.OrderLineItems;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 
 import java.time.LocalDateTime;
 import java.util.List;
